refactor and restructure
