+++ mobile/lib/pages/profile.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class ProfilePage extends StatefulWidget {
+  ProfilePage();
+
+  @override
+  _ProfilePageState createState() => _ProfilePageState();
+}
+
+class _ProfilePageState extends State<ProfilePage> {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      color: Colors.blue,
+    );
+  }
+}
