+++ mobile/lib/pages/search_movie.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class SearchMoviePage extends StatefulWidget {
+  SearchMoviePage({Key key}) : super(key: key);
+
+  @override
+  _SearchMoviePageState createState() => _SearchMoviePageState();
+}
+
+class _SearchMoviePageState extends State<SearchMoviePage> {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+       color: Colors.red,
+    );
+  }
+}
