--- mobile/lib/screens/profile.dart
@@ -1,17 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ProfileScreen extends StatefulWidget {
-  ProfileScreen();
-
-  @override
-  _ProfileScreenState createState() => _ProfileScreenState();
-}
-
-class _ProfileScreenState extends State<ProfileScreen> {
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      color: Colors.blue,
-    );
-  }
-}
