--- mobile/lib/screens/search_movie.dart
@@ -1,17 +1,0 @@
-import 'package:flutter/material.dart';
-
-class SearchMovieScreen extends StatefulWidget {
-  SearchMovieScreen({Key key}) : super(key: key);
-
-  @override
-  _SearchMovieScreenState createState() => _SearchMovieScreenState();
-}
-
-class _SearchMovieScreenState extends State<SearchMovieScreen> {
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-       color: Colors.red,
-    );
-  }
-}
